+++ Assets/Scripts/UI/HealthBar.cs
@@ -5,13 +5,13 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
 
-    public void SetCurrentHealth(int health)
+    public void SetCurrentHealth(float health)
     {
         slider.value = health;
     }
